:new: TIPI-1391: support alternative operators representations or, and,not...
Adds hidden predefined macros for all altnernative tokens in C++ mode,
as in C++ those are valid tokens but not in C this only defines them
in C++ mode independently of the standard version.

Despite C++20 first fully dropped the compatibility header <ciso646>, g++
supports the alternative token independently of the C++ standard version
and the C <iso646.h> doesn't define any of those as macros when the
compiler is in __cplusplus mode.

The rationale to behave equivalently on gcc/clang and MSVC is because we
are targetting standard conformance as a future-proof implementation
choice.

Indeed while MSVC only supports the alternative operators in `/permissive-`
standard conformance mode, Visual Studio default settings for new C++
projects have `/permissive-` ON, therefore we do not implement a specific
MSVC behaviour as it seems bound to disappear in favor of the C++ standard
behaviour.

Fix tipi-build/specs-cmake-re#144